+++ tests/doc/test_rst.py
@@ -15,6 +15,7 @@ def test_amp():
 
     from torchjd.aggregation import UPGrad
     from torchjd.autojac import mtl_backward
+    from torchjd.utils import jac_to_grad
 
     shared_module = Sequential(Linear(10, 5), ReLU(), Linear(5, 3), ReLU())
     task1_module = Linear(3, 1)
@@ -42,9 +43,11 @@ def test_amp():
             loss2 = loss_fn(output2, target2)
 
         scaled_losses = scaler.scale([loss1, loss2])
-        mtl_backward(losses=scaled_losses, features=features, aggregator=aggregator)
+        mtl_backward(losses=scaled_losses, features=features)
+        jac_to_grad(shared_module.parameters(), aggregator)
         scaler.step(optimizer)
         scaler.update()
+        optimizer.zero_grad()
         optimizer.zero_grad()
 
 
@@ -55,6 +58,7 @@ def test_basic_usage():
 
     from torchjd import autojac
     from torchjd.aggregation import UPGrad
+    from torchjd.utils import jac_to_grad
 
     model = Sequential(Linear(10, 5), ReLU(), Linear(5, 2))
     optimizer = SGD(model.parameters(), lr=0.1)
@@ -69,7 +73,8 @@ def test_basic_usage():
     loss1 = loss_fn(output[:, 0], target1)
     loss2 = loss_fn(output[:, 1], target2)
 
-    autojac.backward([loss1, loss2], aggregator)
+    autojac.backward([loss1, loss2])
+    jac_to_grad(model.parameters(), aggregator)
     optimizer.step()
     optimizer.zero_grad()
 
@@ -118,6 +123,7 @@ def test_iwmtl():
         losses.backward(weights)
         optimizer.step()
         optimizer.zero_grad()
+        optimizer.zero_grad()
 
 
 def test_iwrm():
@@ -141,6 +147,7 @@ def test_iwrm():
             loss.backward()
             optimizer.step()
             optimizer.zero_grad()
+            optimizer.zero_grad()
 
     def test_autojac():
         import torch
@@ -149,6 +156,7 @@ def test_iwrm():
 
         from torchjd.aggregation import UPGrad
         from torchjd.autojac import backward
+        from torchjd.utils import jac_to_grad
 
         X = torch.randn(8, 16, 10)
         Y = torch.randn(8, 16)
@@ -163,7 +171,8 @@ def test_iwrm():
         for x, y in zip(X, Y):
             y_hat = model(x).squeeze(dim=1)  # shape: [16]
             losses = loss_fn(y_hat, y)  # shape: [16]
-            backward(losses, aggregator)
+            backward(losses)
+            jac_to_grad(model.parameters(), aggregator)
             optimizer.step()
             optimizer.zero_grad()
 
@@ -194,6 +203,7 @@ def test_iwrm():
             losses.backward(weights)
             optimizer.step()
             optimizer.zero_grad()
+            optimizer.zero_grad()
 
     test_autograd()
     test_autojac()
@@ -220,6 +230,7 @@ def test_lightning_integration():
 
     from torchjd.aggregation import UPGrad
     from torchjd.autojac import mtl_backward
+    from torchjd.utils import jac_to_grad
 
     class Model(LightningModule):
         def __init__(self):
@@ -240,7 +251,9 @@ def test_lightning_integration():
             loss2 = mse_loss(output2, target2)
 
             opt = self.optimizers()
-            mtl_backward(losses=[loss1, loss2], features=features, aggregator=UPGrad())
+
+            mtl_backward(losses=[loss1, loss2], features=features)
+            jac_to_grad(self.feature_extractor.parameters(), UPGrad())
             opt.step()
             opt.zero_grad()
 
@@ -275,6 +288,7 @@ def test_monitoring():
 
     from torchjd.aggregation import UPGrad
     from torchjd.autojac import mtl_backward
+    from torchjd.utils import jac_to_grad
 
     def print_weights(_, __, weights: torch.Tensor) -> None:
         """Prints the extracted weights."""
@@ -314,7 +328,8 @@ def test_monitoring():
         loss1 = loss_fn(output1, target1)
         loss2 = loss_fn(output2, target2)
 
-        mtl_backward(losses=[loss1, loss2], features=features, aggregator=aggregator)
+        mtl_backward(losses=[loss1, loss2], features=features)
+        jac_to_grad(shared_module.parameters(), aggregator)
         optimizer.step()
         optimizer.zero_grad()
 
@@ -326,6 +341,7 @@ def test_mtl():
 
     from torchjd.aggregation import UPGrad
     from torchjd.autojac import mtl_backward
+    from torchjd.utils import jac_to_grad
 
     shared_module = Sequential(Linear(10, 5), ReLU(), Linear(5, 3), ReLU())
     task1_module = Linear(3, 1)
@@ -351,7 +367,8 @@ def test_mtl():
         loss1 = loss_fn(output1, target1)
         loss2 = loss_fn(output2, target2)
 
-        mtl_backward(losses=[loss1, loss2], features=features, aggregator=aggregator)
+        mtl_backward(losses=[loss1, loss2], features=features)
+        jac_to_grad(shared_module.parameters(), aggregator)
         optimizer.step()
         optimizer.zero_grad()
 
@@ -387,6 +404,7 @@ def test_partial_jd():
         losses.backward(weights)
         optimizer.step()
         optimizer.zero_grad()
+        optimizer.zero_grad()
 
 
 def test_rnn():
@@ -396,6 +414,7 @@ def test_rnn():
 
     from torchjd.aggregation import UPGrad
     from torchjd.autojac import backward
+    from torchjd.utils import jac_to_grad
 
     rnn = RNN(input_size=10, hidden_size=20, num_layers=2)
     optimizer = SGD(rnn.parameters(), lr=0.1)
@@ -408,6 +427,7 @@ def test_rnn():
         output, _ = rnn(input)  # output is of shape [5, 3, 20].
         losses = ((output - target) ** 2).mean(dim=[1, 2])  # 1 loss per sequence element.
 
-        backward(losses, aggregator, parallel_chunk_size=1)
+        backward(losses, parallel_chunk_size=1)
+        jac_to_grad(rnn.parameters(), aggregator)
         optimizer.step()
         optimizer.zero_grad()